Add update feature to Task
Tasks cannot be updated after creation, limiting flexibility
when user makes mistakes. Currently, users are forced to
delete and add tasks manually to edit them.

Let's add an update() method to the Task class and implement
it in the Task subclasses (Todo, Deadline and Event). This
allows the tasks to be updated.

* Todo supports updating description.
* Deadline supports updating any of the following:
Description, due date.
* Event supports updating any of the following:
Description, start date and time, end date and time.

This design keeps the logic encapsulated within each Task
subclass. This ensures the proper formats are followed.